+++ api/src/main/java/com/yanny/ali/api/IClientUtils.java
@@ -3,12 +3,14 @@ package com.yanny.ali.api;
 import com.mojang.datafixers.util.Pair;
 import net.minecraft.advancements.critereon.EntitySubPredicate;
 import net.minecraft.advancements.critereon.ItemSubPredicate;
-import net.minecraft.core.HolderLookup;
 import net.minecraft.core.Holder;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.component.DataComponentType;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.consume_effects.ConsumeEffect;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.LootTable;
@@ -26,11 +28,19 @@ public interface IClientUtils extends ICommonUtils {
                                                  List<LootItemFunction> functions, List<LootItemCondition> conditions);
 
     <T extends LootItemCondition> List<Component> getConditionTooltip(IClientUtils utils, int pad, T condition);
+
     <T extends LootItemFunction> List<Component> getFunctionTooltip(IClientUtils utils, int pad, T function);
+
     <T extends ItemSubPredicate> List<Component> getItemSubPredicateTooltip(IClientUtils utils, int pad, ItemSubPredicate.Type<?> type, T predicate);
+
     <T extends EntitySubPredicate> List<Component> getEntitySubPredicateTooltip(IClientUtils utils, int pad, T predicate);
 
+    List<Component> getDataComponentTypeTooltip(IClientUtils utils, int pad, DataComponentType<?> type, Object value);
+
+    <T extends ConsumeEffect> List<Component> getConsumeEffectTooltip(IClientUtils utils, int pad, T effect);
+
     <T extends LootItemFunction> void applyCountModifier(IClientUtils utils, T function, Map<Holder<Enchantment>, Map<Integer, RangeValue>> count);
+
     <T extends LootItemCondition> void applyChanceModifier(IClientUtils utils, T condition, Map<Holder<Enchantment>, Map<Integer, RangeValue>> chance);
 
     <T extends LootItemFunction> ItemStack applyItemStackModifier(IClientUtils utils, T function, ItemStack itemStack);